Wait for OIDC 'ready' event as Promise
It was suggested at https://github.com/okta/okta-oidc-js/issues/68
that the "unknown authentication strategy 'oidc'" error could be
due to a race condition, and that one should wait for the "ready"
event from ExpressOIDC. This change converts the "ready" event into
a Promise and updates the handler to wait for this promise before
proxying the API Gatewway event.

With this change, the "unknown authentication strategy" error goes
away, but an unspecified internal error takes its place.